+++ contact-forms-anti-spam/includes/functions.php
@@ -541,18 +541,6 @@ function maspik_get_settings($data_name, $type = '', $table_var = 'new'){
             }
         }
 
-    //make default value for spam log
-        function set_spam_log_limit_default(){
-            $log_active = maspik_get_settings("maspik_Store_log");
-            $limit_value = maspik_get_settings("spam_log_limit");
-
-            if( ($log_active == "yes") && ($limit_value == '') ){
-
-                maspik_save_settings('spam_log_limit', 1000);
-
-            }
-
-        }
 
     //runs transfer - list of rows to be transfered
         function maspik_run_transfer(){
@@ -714,10 +702,8 @@ function efas_get_browser_name($user_agent){
 function maspik_limit_log_size() {
     global $wpdb;
 
-    $max_logs = maspik_get_settings('spam_log_limit');
-    if (empty($max_logs)) {
-        $max_logs = 1000; // Default value if not set
-    }
+    $max_logs = maspik_get_settings('spam_log_limit') ? maspik_get_settings('spam_log_limit') : 1000;
+
 
     $table = maspik_get_logtable();
 
@@ -1275,6 +1261,218 @@ function efas_array_of_countries(){
 	);
 } 
 
+function maspik_countries_code_for_phone_number(){
+    return array(
+        'none' => 'None (User will enter the phone number with the country code)',
+        'AL' => 'Albania (355)',
+        'DZ' => 'Algeria (213)',
+        'AS' => 'American Samoa (1684)',
+        'AD' => 'Andorra (376)',
+        'AO' => 'Angola (244)',
+        'AI' => 'Anguilla (1264)',
+        'AG' => 'Antigua And Barbuda (1268)',
+        'AR' => 'Argentina (54)',
+          'AM' => 'Armenia (374)',
+          'AW' => 'Aruba (297)',
+          'AU' => 'Australia (61)',
+          'AT' => 'Austria (43)',
+          'AZ' => 'Azerbaijan (994)',
+          'BS' => 'Bahamas (1242)',
+          'BH' => 'Bahrain (973)',
+          'BD' => 'Bangladesh (880)',
+          'BB' => 'Barbados (1246)',
+          'BY' => 'Belarus (375)',
+          'BE' => 'Belgium (32)',
+          'BZ' => 'Belize (501)',
+          'BJ' => 'Benin (229)',
+          'BM' => 'Bermuda (1441)',
+          'BT' => 'Bhutan (975)',
+          'BO' => 'Bolivia (591)',
+          'BA' => 'Bosnia And Herzegovina (387)',
+          'BW' => 'Botswana (267)',
+          'BR' => 'Brazil (55)',
+          'IO' => 'British Indian Ocean Territory (246)',
+          'BN' => 'Brunei (673)',
+          'BG' => 'Bulgaria (359)',
+          'BF' => 'Burkina Faso (226)',
+          'BI' => 'Burundi (257)',
+          'KH' => 'Cambodia (855)',
+          'CM' => 'Cameroon (237)',
+          'CA' => 'Canada (1)',
+          'CV' => 'Cape Verde (238)',
+          'KY' => 'Cayman Islands (1345)',
+          'CF' => 'Central African Republic (236)',
+          'TD' => 'Chad (235)',
+          'CL' => 'Chile (56)',
+          'CN' => 'China (86)',
+          'CO' => 'Colombia (57)',
+          'CG' => 'Congo (242)',
+          'CK' => 'Cook Islands (682)',
+          'CR' => 'Costa Rica (506)',
+          'CI' => 'Cote D\'ivoire (225)',
+          'HR' => 'Croatia (385)',
+          'CU' => 'Cuba (53)',
+          'CY' => 'Cyprus (357)',
+          'CZ' => 'Czech Republic (420)',
+          'CD' => 'Democratic Republic of the Congo (243)',
+          'DK' => 'Denmark (45)',
+          'DJ' => 'Djibouti (253)',
+          'DM' => 'Dominica (1767)',
+          'DO' => 'Dominican Republic (1809)',
+          'EC' => 'Ecuador (593)',
+          'EG' => 'Egypt (20)',
+          'SV' => 'El Salvador (503)',
+          'GQ' => 'Equatorial Guinea (240)',
+          'ER' => 'Eritrea (291)',
+          'EE' => 'Estonia (372)',
+          'ET' => 'Ethiopia (251)',
+          'FO' => 'Faroe Islands (298)',
+          'FM' => 'Federated States Of Micronesia (691)',
+          'FJ' => 'Fiji (679)',
+          'FI' => 'Finland (358)',
+          'FR' => 'France (33)',
+          'GF' => 'French Guiana (594)',
+          'PF' => 'French Polynesia (689)',
+          'GA' => 'Gabon (241)',
+          'GM' => 'Gambia (220)',
+          'GE' => 'Georgia (995)',
+          'DE' => 'Germany (49)',
+          'GH' => 'Ghana (233)',
+          'GI' => 'Gibraltar (350)',
+          'GR' => 'Greece (30)',
+          'GL' => 'Greenland (299)',
+          'GD' => 'Grenada (1473)',
+          'GP' => 'Guadeloupe (590)',
+          'GU' => 'Guam (1671)',
+          'GT' => 'Guatemala (502)',
+          'GN' => 'Guinea (224)',
+          'GW' => 'Guinea Bissau (245)',
+          'GY' => 'Guyana (592)',
+          'HT' => 'Haiti (509)',
+          'HN' => 'Honduras (504)',
+          'HK' => 'Hong Kong (852)',
+          'HU' => 'Hungary (36)',
+          'IS' => 'Iceland (354)',
+          'IN' => 'India (91)',
+          'ID' => 'Indonesia (62)',
+          'IR' => 'Iran (98)',
+          'IE' => 'Ireland (353)',
+          'IL' => 'Israel (972)',
+          'IM' => 'Isle of Man (44)',
+          'IT' => 'Italy (39)',
+          'JM' => 'Jamaica (1876)',
+          'JP' => 'Japan (81)',
+          'JO' => 'Jordan (962)',
+          'KZ' => 'Kazakhstan (7)',
+          'KE' => 'Kenya (254)',
+          'KW' => 'Kuwait (965)',
+          'KG' => 'Kyrgyzstan (996)',
+          'LA' => 'Laos (856)',
+          'LV' => 'Latvia (371)',
+          'LB' => 'Lebanon (961)',
+          'LS' => 'Lesotho (266)',
+          'LY' => 'Libyan Arab Jamahiriya (218)',
+          'LI' => 'Liechtenstein (423)',
+          'LT' => 'Lithuania (370)',
+          'LU' => 'Luxembourg (352)',
+          'MK' => 'Macedonia (389)',
+          'MG' => 'Madagascar (261)',
+          'MW' => 'Malawi (265)',
+          'MY' => 'Malaysia (60)',
+          'MV' => 'Maldives (960)',
+          'ML' => 'Mali (223)',
+          'MT' => 'Malta (356)',
+          'MQ' => 'Martinique (596)',
+          'MR' => 'Mauritania (222)',
+          'MU' => 'Mauritius (230)',
+          'MX' => 'Mexico (52)',
+          'MC' => 'Monaco (377)',
+          'MN' => 'Mongolia (976)',
+          'ME' => 'Montenegro (382)',
+          'MA' => 'Morocco (212)',
+          'MZ' => 'Mozambique (258)',
+          'MM' => 'Myanmar (95)',
+          'NA' => 'Namibia (264)',
+          'NP' => 'Nepal (977)',
+          'NL' => 'Netherlands (31)',
+          'AN' => 'Netherlands Antilles (599)',
+          'NC' => 'New Caledonia (687)',
+          'NZ' => 'New Zealand (64)',
+          'NI' => 'Nicaragua (505)',
+          'NE' => 'Niger (227)',
+          'NG' => 'Nigeria (234)',
+          'NF' => 'Norfolk Island (672)',
+          'MP' => 'Northern Mariana Islands (1670)',
+          'NO' => 'Norway (47)',
+          'OM' => 'Oman (968)',
+          'PK' => 'Pakistan (92)',
+          'PW' => 'Palau (680)',
+          'PA' => 'Panama (507)',
+          'PG' => 'Papua New Guinea (675)',
+          'PY' => 'Paraguay (595)',
+          'PE' => 'Peru (51)',
+          'PH' => 'Philippines (63)',
+          'PL' => 'Poland (48)',
+          'PT' => 'Portugal (351)   ',
+          'PR' => 'Puerto Rico (1787)',
+          'QA' => 'Qatar (974)',
+          'MD' => 'Republic Of Moldova (373)',
+          'RE' => 'Reunion (262)',
+          'RO' => 'Romania (40)',
+          'RU' => 'Russia (7)',
+          'RW' => 'Rwanda (250)',
+          'KN' => 'Saint Kitts And Nevis (1869)',
+          'LC' => 'Saint Lucia (1758)',
+          'VC' => 'Saint Vincent And The Grenadines (1784)',
+          'WS' => 'Samoa (685)',
+          'SM' => 'San Marino (378)',
+          'ST' => 'Sao Tome And Principe (239)',
+          'SA' => 'Saudi Arabia (966)',
+          'SN' => 'Senegal (221)',
+          'RS' => 'Serbia (381)',
+          'SC' => 'Seychelles (248)',
+          'SG' => 'Singapore (65)',
+          'SK' => 'Slovakia (421)',
+          'SI' => 'Slovenia (386)',
+          'SB' => 'Solomon Islands (677)',
+          'ZA' => 'South Africa (27)',
+          'KR' => 'South Korea (82)',
+          'ES' => 'Spain (34)',
+          'LK' => 'Sri Lanka (94)',
+          'SD' => 'Sudan (249)',
+          'SR' => 'Suriname (597)',
+          'SZ' => 'Swaziland (268)',
+          'SE' => 'Sweden (46)',
+          'CH' => 'Switzerland (41)',
+          'SY' => 'Syrian Arab Republic (963)',
+          'TW' => 'Taiwan (886)',
+          'TJ' => 'Tajikistan (992)',
+          'TZ' => 'Tanzania (255)',
+          'TH' => 'Thailand (66)',
+          'TG' => 'Togo (228)',
+          'TO' => 'Tonga (676)',
+          'TT' => 'Trinidad And Tobago (1868)',
+          'TN' => 'Tunisia (216)',
+          'TR' => 'Turkey (90)',
+          'TM' => 'Turkmenistan (993)',
+          'UG' => 'Uganda (256)',
+          'UA' => 'Ukraine (380)',
+          'AE' => 'United Arab Emirates (971)',
+          'GB' => 'United Kingdom (44)',
+          'US' => 'United States (1)',
+          'UY' => 'Uruguay (598)',
+          'UZ' => 'Uzbekistan (998)',
+          'VU' => 'Vanuatu (678)',
+          'VE' => 'Venezuela (58)',
+          'VN' => 'Vietnam (84)',
+          'VG' => 'Virgin Islands British (1284)',
+          'VI' => 'Virgin Islands U.S. (1340)',
+          'YE' => 'Yemen (967)',
+          'ZM' => 'Zambia (260)',
+          'ZW' => 'Zimbabwe (263)'
+      );
+  } 
+  
 function maspik_is_plugin_active( $plugin ) {
 	require_once( ABSPATH . '/wp-admin/includes/plugin.php' );
 	return in_array( $plugin, (array) get_option( 'active_plugins', array() ), true ) || is_plugin_active_for_network( $plugin );
@@ -1514,7 +1712,7 @@ function cfas_refresh_api($type = 'regular') {
         return;
     }
 
-    $private_file_id = maspik_get_settings('private_file_id');
+    $private_file_id = (int)maspik_get_settings('private_file_id');
     $domain = isset($_SERVER['SERVER_NAME']) ? sanitize_text_field(wp_unslash($_SERVER['SERVER_NAME'])) : '';
 
     // Initialize $file as an empty array